backend/requirements.txt
aider-install==0.1.3
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.16.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
apprise==1.9.3
asgi-lifespan==2.1.0
asgiref==3.8.1
asyncpg==0.30.0
attrs==25.3.0
azure-ai-inference==1.0.0b9
azure-ai-ml==1.26.3
azure-common==1.1.28
azure-core==1.33.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-identity==1.21.0
azure-mgmt-core==1.5.0
azure-monitor-opentelemetry==1.6.8
azure-monitor-opentelemetry-exporter==1.0.0b36
azure-storage-blob==12.25.1
azure-storage-file-datalake==12.20.0
azure-storage-file-share==12.21.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.2
coolname==2.2.0
cryptography==44.0.2
cycler==0.12.1
dataclasses-json==0.6.7
dateparser==1.2.1
Deprecated==1.2.18
distro==1.9.0
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
docker==7.1.0
exceptiongroup==1.3.0
faiss-cpu==1.11.0
fastapi==0.115.12
filelock==3.18.0
fixedint==0.1.6
Flask==3.1.0
fonttools==4.57.0
fpdf==1.7.2
frozendict==2.4.6
frozenlist==1.6.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
googletrans==4.0.0rc1
graphviz==0.20.3
greenlet==3.2.1
griffe==1.7.3
h11==0.16.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httpx==0.25.2
httpx-sse==0.4.0
huggingface-hub==0.31.2
humanize==4.12.3
hyperframe==5.2.0
idna==2.10
importlib_metadata==8.6.1
iniconfig==2.1.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
langchain-huggingface==0.2.0
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
langgraph==0.4.5
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.69
langsmith==0.3.42
lxml==5.3.2
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
msal==1.32.0
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.4.3
multitasking==0.0.11
mypy_extensions==1.1.0
narwhals==1.35.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.5
oauthlib==3.2.2
ollama==0.1.6
openai==1.78.1
opentelemetry-api==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-dbapi==0.52b1
opentelemetry-instrumentation-django==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-instrumentation-flask==0.52b1
opentelemetry-instrumentation-psycopg2==0.52b1
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-urllib==0.52b1
opentelemetry-instrumentation-urllib3==0.52b1
opentelemetry-instrumentation-wsgi==0.52b1
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
orjson==3.10.18
ormsgpack==1.9.1
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
peewee==3.17.9
pillow==11.2.1
platformdirs==4.3.7
plotly==6.0.1
pluggy==1.6.0
prefect==3.4.4
prometheus_client==0.22.0
propcache==0.3.1
protobuf==5.29.4
psutil==6.1.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyarrow==19.0.1
pycparser==2.22
pydantic==2.11.4
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydash==8.0.5
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pypdf==5.5.0
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-slugify==8.0.4
python-socks==2.7.1
pytrends==4.9.2
pytz==2025.2
pywin32==310
PyYAML==6.0.2
readchar==4.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==14.0.0
rpds-py==0.24.0
ruamel.yaml==0.18.12
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
selenium==4.31.0
sentence-transformers==4.1.0
setuptools==80.7.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.40
sqlparse==0.5.3
starlette==0.46.2
streamlit==1.44.1
strictyaml==1.7.3
sympy==1.14.0
tenacity==9.1.2
text-unidecode==1.3
textblob==0.19.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.3
trio==0.30.0
trio-websocket==0.12.2
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
urllib3==2.4.0
uv==0.7.2
uvicorn==0.22.0
watchdog==6.0.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
whenever==0.8.4
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.0
yfinance==0.2.55
zipp==3.21.0
zstandard==0.23.0
